backend/requirements.txt
# Yadro Post Backend
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
httpx>=0.27.0
apscheduler==3.10.4
anthropic>=0.18.0
python-telegram-bot==21.0.1
vk-api==11.9.9
python-multipart==0.0.6
aiofiles==23.2.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
redis==5.0.1
